+++ api/cases/validations/input.js
@@ -28,36 +28,10 @@ const CaseVerifyPayload = Joi.object().keys({
     verified_comment: Joi.string().allow('', null).optional()
 })
 
-const CaseTransferPayload = Joi.object().keys({
-    transfer_to_unit_id: Joi.string().required(),
-    transfer_to_unit_name: Joi.string().required(),
-    transfer_comment: Joi.string().allow('', null).optional().default(null),
-})
-
-const CaseTransferActionPayload = Joi.object().keys({
-    transfer_comment: Joi.string().allow('', null).optional().default(null),
-})
-
 const CaseParamsValidations = {
     params: {
         id: Joi.string().required()
     }
-}
-
-const TransferActionParamsValidations = {
-    params: {
-        id: Joi.string().required(),
-        transferId: Joi.string().required(),
-        action: Joi.string().valid('approve','decline','abort').required(),
-    }
-}
-
-const TransferCaseListParamValidations = {
-    params: {
-        type: Joi.string().valid('in','out').required(),
-    },
-    options: validateOptions.options,
-    failAction: validateOptions.failAction
 }
 
 // --------------------------------------------------
@@ -174,20 +148,6 @@ const CaseVerifyPayloadValidations = Object.assign({
     failAction: validateOptions.failAction
 }, CaseParamsValidations)
 
-const CaseTransferPayloadValidations = Object.assign({
-    payload: CaseTransferPayload,
-    headers: HeadersPayLoad,
-    options: validateOptions.options,
-    failAction: validateOptions.failAction
-}, CaseParamsValidations)
-
-const CaseTransferActPayloadValidations = Object.assign({
-    payload: CaseTransferActionPayload,
-    headers: HeadersPayLoad,
-    options: validateOptions.options,
-    failAction: validateOptions.failAction
-}, TransferActionParamsValidations)
-
 module.exports = {
     CaseParamsValidations,
     CaseQueryValidations,
@@ -196,8 +156,5 @@ module.exports = {
     CaseDeletePayloadValidations,
     CaseImportPayloadValidations,
     caseSchemaValidation,
-    CaseVerifyPayloadValidations,
-    CaseTransferPayloadValidations,
-    CaseTransferActPayloadValidations,
-    TransferCaseListParamValidations,
+    CaseVerifyPayloadValidations
 }
