+++ api/cases_transfers/package.json
@@ -1,0 +1,4 @@
+{
+    "name" : "cases_transfers",
+    "version": "1.0.0"
+}
